feat(deps): migrate xanzy/go-gitlab → gitlab.com/gitlab-org/api/client-go
xanzy/go-gitlab has been archived since 2024. The official module is
gitlab.com/gitlab-org/api/client-go v1.46.0.

Breaking API changes handled:
- Project/group/deploy-key IDs are now int64 (were int)
- FileAction() helper removed; CommitActionOptions.Action is now
  *FileActionValue — use &actionValue directly (or gitlab.Ptr())
- FluxDeployKeyID field type updated to int64; cfg.FluxCDDeployKey cast
  to int64 at the call site in main.go

Also re-enables SA1019 in golangci-lint (the exclusion comment tracked
this migration; no other deprecated symbols remain).

+++ cmd/server/main.go
@@ -107,7 +107,7 @@ func run() error {
 		ProjectID:             cfg.GitLabProjectID,
 		ArgoCDGroupID:         cfg.GitLabGroupID,
 		ArgoCDPath:            cfg.GitLabArgoCDPath,
-		FluxDeployKeyID:       cfg.FluxCDDeployKey,
+		FluxDeployKeyID:       int64(cfg.FluxCDDeployKey),
 		DomainPreprod:         cfg.BaseDomainPreprod,
 		DomainProd:            cfg.BaseDomainProd,
 		VaultAddr:             cfg.VaultAddr,
